flask_server/requirements.txt
Flask~=3.0.0
neo4j~=5.15.0
psycopg2~=2.9.9
python-dotenv~=1.0.0